move tests that has external dependencies (taxonomy api)  to 'integration_tests' folder
