--- tests/unit_tests/test_platsannonser.py
@@ -1,11 +1,0 @@
-import sokannonser.repository.platsannonser
-
-# Todo:  test _extract_concept_from_concepts
-
-# todo: test _format_ad
-
-# todo: get_not_found_logo_file
-
-# todo: test _modify_results
-
-# todo: test find_agg_and_delete
